requirements.txt
ping3~=4.0.4
numpy~=1.26.0
psutil~=5.9.5
pika~=1.3.2
requests~=2.31.0
iperf3~=0.1.11
kafka-python~=2.0.2
prometheus_client~=0.20.0
flask~=3.0.3